fix type mismatch

+++ src/zippylog/client.cpp
@@ -300,7 +300,7 @@ bool Client::HandleSubscriptionResponse(Envelope &e, SubscriptionStart &start, v
 
     SubscriptionCallback cb = iter->second.cb;
 
-    for (size_t i = 1; i < e.MessageCount(); i++) {
+    for (int i = 1; i < e.MessageCount(); i++) {
         switch (e.MessageType(i)) {
             case protocol::StoreChangeBucketAdded::zippylog_enumeration:
                 if (cb.BucketAdded) {
